requirement.txt
beautifulsoup4==4.14.2
openai==2.8.1
requests==2.32.5
python-dotenv==1.2.1